Move CA1036 to its new location - no functional changes
