+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/UnitTests/OverrideMethodsOnComparableTypesTests.Fixer.cs
@@ -1,23 +1,21 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Test.Utilities;
-using Microsoft.CodeAnalysis.UnitTests;
 using Xunit;
 
 namespace Microsoft.ApiDesignGuidelines.Analyzers.UnitTests
 {
-    public class OverrideMethodsOnComparableTypesFixerTests : CodeFixTestBase
+    public partial class OverrideMethodsOnComparableTypesTests
     {
-        protected override DiagnosticAnalyzer GetBasicDiagnosticAnalyzer()
+        protected override CodeFixProvider GetCSharpCodeFixProvider()
         {
-            return new BasicOverrideMethodsOnComparableTypesAnalyzer();
-        }
-
-        protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
-        {
-            return new CSharpOverrideMethodsOnComparableTypesAnalyzer();
+            return new CSharpOverrideMethodsOnComparableTypesFixer();
         }
 
         protected override CodeFixProvider GetBasicCodeFixProvider()
@@ -25,9 +23,228 @@ namespace Microsoft.ApiDesignGuidelines.Analyzers.UnitTests
             return new BasicOverrideMethodsOnComparableTypesFixer();
         }
 
-        protected override CodeFixProvider GetCSharpCodeFixProvider()
+        [Fact]
+        public void CA1036GenerateAllCSharp()
         {
-            return new CSharpOverrideMethodsOnComparableTypesFixer();
+            VerifyCSharpFix(@"
+using System;
+
+public class A : IComparable
+{    
+    public int CompareTo(object obj)
+    {
+        return 1;
+    }
+}
+", @"
+using System;
+
+public class A : IComparable
+{    
+    public int CompareTo(object obj)
+    {
+        return 1;
+    }
+
+    public override bool Equals(object obj)
+    {
+        throw new NotImplementedException();
+    }
+
+    public override int GetHashCode()
+    {
+        throw new NotImplementedException();
+    }
+
+    public static bool operator ==(A left, A right)
+    {
+        throw new NotImplementedException();
+    }
+
+    public static bool operator !=(A left, A right)
+    {
+        throw new NotImplementedException();
+    }
+
+    public static bool operator <(A left, A right)
+    {
+        throw new NotImplementedException();
+    }
+
+    public static bool operator >(A left, A right)
+    {
+        throw new NotImplementedException();
+    }
+}
+");
+        }
+
+        [Fact]
+        public void CA1036GenerateSomeCSharp()
+        {
+            VerifyCSharpFix(@"
+using System;
+
+public class A : IComparable
+{    
+    public override int GetHashCode()
+    {
+        return 1234;
+    }
+
+    public int CompareTo(object obj)
+    {
+        return 1;
+    }
+
+    public static bool operator !=(A objLeft, A objRight)
+    {
+        return true;
+    }
+}
+", @"
+using System;
+
+public class A : IComparable
+{    
+    public override int GetHashCode()
+    {
+        return 1234;
+    }
+
+    public int CompareTo(object obj)
+    {
+        return 1;
+    }
+
+    public static bool operator !=(A objLeft, A objRight)
+    {
+        return true;
+    }
+
+    public override bool Equals(object obj)
+    {
+        throw new NotImplementedException();
+    }
+
+    public static bool operator ==(A left, A right)
+    {
+        throw new NotImplementedException();
+    }
+
+    public static bool operator <(A left, A right)
+    {
+        throw new NotImplementedException();
+    }
+
+    public static bool operator >(A left, A right)
+    {
+        throw new NotImplementedException();
+    }
+}
+");
+        }
+
+        [Fact]
+        public void CA1036GenerateAllVisualBasic()
+        {
+            VerifyBasicFix(@"
+Imports System
+
+Public Class A : Implements IComparable
+
+    Public Function CompareTo(obj As Object) As Integer
+        Return 1
+    End Function
+
+End Class", @"
+Imports System
+
+Public Class A : Implements IComparable
+
+    Public Function CompareTo(obj As Object) As Integer
+        Return 1
+    End Function
+
+    Public Overrides Function Equals(obj As Object) As Boolean
+        Throw New NotImplementedException()
+    End Function
+
+    Public Overrides Function GetHashCode() As Integer
+        Throw New NotImplementedException()
+    End Function
+
+    Public Shared Operator =(left As A, right As A) As Boolean
+        Throw New NotImplementedException()
+    End Operator
+
+    Public Shared Operator <>(left As A, right As A) As Boolean
+        Throw New NotImplementedException()
+    End Operator
+
+    Public Shared Operator <(left As A, right As A) As Boolean
+        Throw New NotImplementedException()
+    End Operator
+
+    Public Shared Operator >(left As A, right As A) As Boolean
+        Throw New NotImplementedException()
+    End Operator
+End Class");
+        }
+
+        [Fact]
+        public void CA1036GenerateSomeVisualBasic()
+        {
+            VerifyBasicFix(@"
+Imports System
+
+Public Class A : Implements IComparable
+
+    Public Overrides Function GetHashCode() As Integer
+        Return 1234
+    End Function
+
+    Public Shared Operator <(objLeft As A, objRight As A) As Boolean
+        Return True
+    End Operator
+
+    Public Function CompareTo(obj As Object) As Integer
+        Return 1
+    End Function
+
+End Class", @"
+Imports System
+
+Public Class A : Implements IComparable
+
+    Public Overrides Function GetHashCode() As Integer
+        Return 1234
+    End Function
+
+    Public Shared Operator <(objLeft As A, objRight As A) As Boolean
+        Return True
+    End Operator
+
+    Public Function CompareTo(obj As Object) As Integer
+        Return 1
+    End Function
+
+    Public Overrides Function Equals(obj As Object) As Boolean
+        Throw New NotImplementedException()
+    End Function
+
+    Public Shared Operator =(left As A, right As A) As Boolean
+        Throw New NotImplementedException()
+    End Operator
+
+    Public Shared Operator <>(left As A, right As A) As Boolean
+        Throw New NotImplementedException()
+    End Operator
+
+    Public Shared Operator >(left As A, right As A) As Boolean
+        Throw New NotImplementedException()
+    End Operator
+End Class");
         }
     }
 }
